Create bsa-glaac-ca-ab506-certificates.ipynb
Starting a Jupiter Notebook for analyzing AB 506 data for Greater Los Angeles Area Council of the Boy Scouts of America.

First uploading the files to S. Eventually Amazon Textract will be used to amaze the documents, but for Textract to use synchronous analyzing (the simpler setup) it only works for PDFs with 1 page only, so I am using PyPDF2 to detect if there are multiple pages (in this case the disclosure PDFs are multiple pages and we only need to analyze Page 2, the signature page) and then in those instances saving only Page 2 to S3.